do not install conjure if it's already installed

+++ scripts/install-colab.sh
@@ -11,13 +11,20 @@ echo "Installing Conjure version ${CONJURE_VERSION} and Conjure Notebook version
 # remove the sample-data directory that google colab creates, we don't need them
 rm -rf sample_data
 
-# download latest release from Github
-wget --quiet --no-check-certificate -c https://github.com/conjure-cp/conjure/releases/download/${CONJURE_VERSION}/conjure-${CONJURE_VERSION}-linux-with-solvers.zip
-unzip -q -o conjure-${CONJURE_VERSION}-linux-with-solvers.zip
-cp -r conjure-${CONJURE_VERSION}-linux-with-solvers/* /usr/local/bin
+conjure_installed=$((conjure --version 2> /dev/null) | head -n2 | tail -n1)
 
-# we don't need to keep these around any more
-rm -rf conjure-${CONJURE_VERSION}-linux-with-solvers conjure-${CONJURE_VERSION}-linux-with-solvers.zip
+if [[ $conjure_installed == "Release version 2.4.1" ]]; then
+    echo "Conjure is already installed."
+else
+    echo "Downloading."
+    # download the release from Github
+    wget --quiet --no-check-certificate -c https://github.com/conjure-cp/conjure/releases/download/${CONJURE_VERSION}/conjure-${CONJURE_VERSION}-linux-with-solvers.zip
+    unzip -q -o conjure-${CONJURE_VERSION}-linux-with-solvers.zip
+    cp -r conjure-${CONJURE_VERSION}-linux-with-solvers/* /usr/local/bin
+
+    # we don't need to keep these around any more
+    rm -rf conjure-${CONJURE_VERSION}-linux-with-solvers conjure-${CONJURE_VERSION}-linux-with-solvers.zip
+fi
 
 # installing the conjure extension
 pip --quiet install git+https://github.com/conjure-cp/conjure-notebook.git@${NOTEBOOK_VERSION} > /dev/null
